Fix minor issues reported by CodeQL

+++ perf/PerfProducer.cc
@@ -160,7 +160,7 @@ void startPerfProducer(const Arguments& args, pulsar::ProducerConfiguration& pro
         limiter = std::make_shared<pulsar::RateLimiter>(args.rate);
     }
 
-    producerList.resize(args.numTopics * args.numProducers);
+    producerList.resize((size_t)args.numTopics * args.numProducers);
     for (int i = 0; i < args.numTopics; i++) {
         std::string topic = (args.numTopics == 1) ? args.topic : args.topic + "-" + std::to_string(i);
         LOG_INFO("Adding " << args.numProducers << " producers on topic " << topic);
